Create XML Assignment 2
The xml contains data about customers and thier orders. The orders can include detail such as the item and the quantity along with the item number and information about the order.

The schema is able to verify the xml file by defining the elements used in the xml file. There are several custom datatypes to support the xml data including OrderIDType, title, saleItemType. These custom datatypes use the standard datatype string but add restrictions.